+++ api/index.js
@@ -14,6 +14,7 @@ app.use(bodyParser.json());
 // Import API Routes
 app.use('/auth', require('./routes/auth'));
 // app.use('/db', require('./routes/db'));
+app.use('/', require('./routes/test'));
 
 // // Error handler
 app.use((err, req, res, next) => {